backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
diskcache==5.6.3
fastapi==0.114.2
h11==0.14.0
httptools==0.6.1
idna==3.9
Jinja2==3.1.4
llama_cpp_python==0.2.90
MarkupSafe==2.1.5
numpy==2.1.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
fastapi>=0.114.2

